+++ CHANGELOG.md
@@ -12,10 +12,14 @@
     # settings.py
     COMPONENTS = {
         "extensions": [
-            "djc_ext_pydantic.PydanticExtension",
+            "djc_pydantic.PydanticExtension",
         ],
     }
     ```
+
+#### Fix
+
+- Make it possible to resolve URLs added by extensions by their names
 
 ## v0.135